Move rest-api/service to core/service
